agent_app/pyproject.toml
[project]
name = "agent-server"
version = "0.1.0"
description = "Migrated multi-agent Genie system from Model Serving to Databricks Apps"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
    "mlflow>=3.10.0",
    "databricks-agents>=1.9.3",
    "databricks-ai-bridge[memory]>=0.17.0",
    "databricks-sdk>=0.84.0",
    "databricks-sql-connector>=4.2.4",
    "databricks-langchain[memory]>=0.12.1",
    "databricks-vectorsearch>=0.63",
    "langgraph>=1.1.0",
    "litellm",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "ddgs>=0.1",
    "ddgs>=9.11.4",
    "sqlglot>=30.1.0",
]

[tool.uv]
prerelease = "allow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.28.0",
    "pytest>=9.0.2",
]

[project.scripts]
quickstart = "scripts.quickstart:main"
start-app = "scripts.start_app:main"
start-server = "agent_server.start_server:main"
agent-evaluate = "agent_server.evaluate_agent:evaluate"
discover-tools = "scripts.discover_tools:main"
